refactor(api): extract dashboard service into modular directory
Split the 700-line monolithic dashboard.service.ts into a structured
directory with dedicated files for types, aggregations, health score,
and cash runway outlier filtering.
