+++ packages/api/src/services/index.ts
@@ -3,7 +3,7 @@ import AccountService from './account.service'
 import AuthService from './auth.service'
 import BudgetService from './budget.service'
 import CategoryService from './category.service'
-import DashboardService from './dashboard.service'
+import DashboardService from './dashboard'
 import DebtService from './debt.service'
 import LoanService from './loan.service'
 import PensionService from './pension.service'
